requirements.txt
asgiref==3.11.1
channels==4.3.2
daphne==4.1.2
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf_spectacular==0.29.0
Pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
